Refactor server variables page
